app/requirements.txt
Django==4.0.4
django-safedelete==1.2.2
django-tenants==3.4.2
psycopg2-binary==2.9.3
